requirements.txt
attrs==20.2.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
idna==2.10
jsonschema==3.2.0
pycparser==2.20
pyrsistent==0.17.3
python-dotenv==0.13.0
python-telegram-bot==12.7
requests==2.24.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
validator-collection==1.4.2
pillow==7.2.0
